added clean webpack plugin + profile aside mostly done + more sass... + each section in its own file + caps first letter
