+++ src/components/Sections/education.js
@@ -1,0 +1,14 @@
+import { h } from 'preact';
+
+/**
+ * @param {Object} props
+ * @returns {JSX}
+ */
+export default (props) => (
+    <section class="">
+        EDU
+        <div class="row">
+            <div class="cd-timeline cd-container verde" />
+        </div>
+    </section>
+);
