+++ src/components/Sections/experience.js
@@ -1,0 +1,14 @@
+import { h } from 'preact';
+
+/**
+ * @param {Object} props
+ * @returns {JSX}
+ */
+export default (props) => (
+    <section class="">
+        XP
+        <div class="row xps-container" id="experiences">
+            <div class="cd-timeline cd-container amarelo" />
+        </div>
+    </section>
+);
