+++ src/components/Sections/index.js
@@ -1,0 +1,13 @@
+import me from './me';
+import skills from './skills';
+import experience from './experience';
+import education from './education';
+import more from './more';
+
+export default {
+    me,
+    skills,
+    experience,
+    education,
+    more,
+};
