+++ src/components/Sections/me.js
@@ -1,0 +1,13 @@
+import { h } from 'preact';
+
+/**
+ * @param {Object} props
+ * @returns {JSX}
+ */
+export default (props) => (
+    <section id="me" class="row">
+        <div class="row">
+            <p class="me-small">I believe in <span class="me-big">Simplicity</span></p>
+        </div>
+    </section>
+);
