+++ src/styles/_all.scss
@@ -1,17 +1,16 @@
-%display-flex {
-  display: flex;
-}
-
-
 #root {
   // Normalize
-  font-family: 'Roboto';
+  font-family: "Roboto";
   box-sizing: border-box;
   // Pirilimpimpim
   @extend %background-color-second;
   @extend %color-third;
-  @extend %display-flex;
   height: 100vh;
+  @extend %display-flex;
+
+  @include onMobile() {
+    flex-wrap: wrap;
+  }
 }
 
 main {
@@ -19,10 +18,6 @@ main {
   height: 100%;
   width: 100%;
   overflow-y: scroll;
-}
-
-#main-content {
-  padding: 0 10px;
 }
 
 /* ************** */
@@ -72,9 +67,9 @@ main {
 
 .skill {
   @extend %display-flex;
-  justify-content: center;
+  @extend %justify-content-center;
   overflow: hidden;
-  margin: 2%; 
+  margin: 2%;
   /* SQUARE */
   position: relative;
   width: 15%;
@@ -85,7 +80,7 @@ main {
 }
 
 .skill-content {
-  flex: 1;
+  @extend %flex-1;
   position: absolute;
   -webkit-box-sizing: border-box; /* Safari Chrome, other WebKit */
   -moz-box-sizing: border-box; /* Firefox, other Gecko */
