+++ src/styles/helpers/_all.scss
@@ -1,0 +1,3 @@
+.-circl {
+    @extend %border-radius-100-percent;
+}
