+++ src/styles/layout/_skills.scss
@@ -1,14 +1,14 @@
 // Skills
-.skills .ttl {
-  text-transform: uppercase;
-  font-weight: bold;
-}
+// .skills .ttl {
+//   text-transform: uppercase;
+//   font-weight: bold;
+// }
 
-section#experience,
-section#education {
-  overflow-x: hidden;
-}
+// section#experience,
+// section#education {
+//   overflow-x: hidden;
+// }
 
-.margem-cima {
-  margin-top: 20vh;
-}
+// .margem-cima {
+//   margin-top: 20vh;
+// }
