+++ src/styles/placeholders/_all.scss
@@ -1,0 +1,10 @@
+// Gathering all in one
+@import "placeholders/border";
+@import "placeholders/flex";
+@import "placeholders/colors";
+@import "placeholders/spaces";
+@import "placeholders/positions";
+
+%content-empty {
+  content: "";
+}
