+++ src/styles/placeholders/_border.scss
@@ -1,0 +1,3 @@
+%border-radius-100-percent {
+    border-radius: 100%;
+}
