+++ src/styles/placeholders/_flex.scss
@@ -1,0 +1,23 @@
+%display-flex {
+  display: flex;
+}
+
+%flex-wrap-wrap {
+  flex-wrap: wrap;
+}
+
+%flex-direction-column {
+  flex-direction: column;
+}
+
+%flex-1 {
+  flex: 1;
+}
+
+%justify-content-center {
+  justify-content: center;
+}
+
+%align-items-center {
+  align-items: center;
+}
