+++ src/styles/placeholders/_positions.scss
@@ -1,0 +1,19 @@
+%position-relative {
+  position: relative;
+}
+
+%position-sticky {
+  position: sticky;
+}
+
+%position-absolute {
+  position: absolute;
+}
+
+%top-0 {
+  top: 0;
+}
+
+%left-0 {
+  left: 0;
+}
