+++ webpack.config.js
@@ -2,6 +2,7 @@ const { join } = require('path');
 const HtmlWebPackPlugin = require('html-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const CleanWebpackPlugin = require('clean-webpack-plugin');
 
 module.exports = (env, argv) => ({
     target: 'web',
@@ -68,6 +69,7 @@ module.exports = (env, argv) => ({
         ],
     },
     plugins: [
+        new CleanWebpackPlugin(),
         new HtmlWebPackPlugin({
             template: './template/index.html',
             filename: './index.html',
